RavenDB-20339 - move classes to dedicated files and add interface
