+++ test/Tests.Infrastructure/ReplicationTestBase.cs
@@ -24,11 +24,9 @@ using Raven.Client.ServerWide.Operations;
 using Raven.Server;
 using Raven.Server.Documents;
 using Raven.Server.Documents.Handlers.Processors.Replication;
-using Raven.Server.Documents.Replication.Stats;
 using Raven.Server.Web;
 using Raven.Server.Web.System;
 using Sparrow.Json;
-using Sparrow.Utils;
 using Xunit;
 using Xunit.Abstractions;
 using Xunit.Sdk;
@@ -41,176 +39,7 @@ namespace Tests.Infrastructure
         {
         }
 
-        public class ReplicationInstance : IDisposable
-        {
-            private readonly DocumentDatabase _database;
-            public readonly string DatabaseName;
-            private ManualResetEventSlim _replicateOnceMre;
-            private bool _replicateOnceInitialized = false;
-
-            protected ReplicationInstance(string databaseName)
-            {
-                DatabaseName = databaseName;
-            }
-
-            public ReplicationInstance(DocumentDatabase database, string databaseName)
-            {
-                _database = database;
-                DatabaseName = databaseName ?? throw new ArgumentNullException(nameof(databaseName));
-            }
-
-            public virtual void Break()
-            {
-                var mre = new ManualResetEventSlim(false);
-                _database.ReplicationLoader.DebugWaitAndRunReplicationOnce = mre;
-            }
-
-            public virtual void Mend()
-            {
-                var mre = _database.ReplicationLoader.DebugWaitAndRunReplicationOnce;
-                Assert.NotNull(mre);
-                _database.ReplicationLoader.DebugWaitAndRunReplicationOnce = null;
-                _database.Configuration.Replication.MaxItemsCount = null;
-                mre.Set();
-            }
-
-            protected virtual void InitializeReplicateOnce()
-            {
-                _database.Configuration.Replication.MaxItemsCount = 1;
-
-                _database.ReplicationLoader.DebugWaitAndRunReplicationOnce ??= new ManualResetEventSlim(true);
-                _replicateOnceMre = _database.ReplicationLoader.DebugWaitAndRunReplicationOnce;
-
-                _replicateOnceInitialized = true;
-            }
-
-            public virtual void ReplicateOnce(string docId)
-            {
-                if(_replicateOnceInitialized == false)
-                    InitializeReplicateOnce();
-
-                WaitForReset(); //wait for server to block and wait
-                _replicateOnceMre.Set(); //let threads pass
-            }
-
-            //wait to reach reset and wait point in server
-            private void WaitForReset(int timeout = 15_000)
-            {
-                var sp = Stopwatch.StartNew();
-                while (sp.ElapsedMilliseconds < timeout)
-                {
-                    if (_replicateOnceMre.IsSet == false)
-                        return;
-
-                    Thread.Sleep(16);
-                }
-
-                throw new TimeoutException();
-            }
-
-            public virtual async Task EnsureNoReplicationLoopAsync()
-            {
-                using (var collector = new LiveReplicationPulsesCollector(_database))
-                {
-                    var etag1 = _database.DocumentsStorage.GenerateNextEtag();
-
-                    await Task.Delay(3000);
-
-                    var etag2 = _database.DocumentsStorage.GenerateNextEtag();
-
-                    Assert.True(etag1 + 1 == etag2, "Replication loop found :(");
-
-                    var groups = collector.Pulses.GetAll().GroupBy(p => p.Direction);
-                    foreach (var group in groups)
-                    {
-                        var key = group.Key;
-                        var count = group.Count();
-                        Assert.True(count < 50, $"{key} seems to be excessive ({count})");
-                    }
-                }
-            }
-
-            public virtual void Dispose()
-            {
-                WaitForReset();
-                _database.ReplicationLoader.DebugWaitAndRunReplicationOnce = null;
-                _database.Configuration.Replication.MaxItemsCount = null;
-                _replicateOnceMre.Set();
-            }
-
-            internal static async ValueTask<ReplicationInstance> GetReplicationInstanceAsync(RavenServer server, string databaseName)
-            {
-                DevelopmentHelper.ShardingToDo(DevelopmentHelper.TeamMember.Stav, DevelopmentHelper.Severity.Normal, "Make this func private when legacy BreakReplication() is removed");
-                return new ReplicationInstance(await server.ServerStore.DatabasesLandlord.TryGetOrCreateResourceStore(databaseName), databaseName);
-            }
-        }
-
-        public class ReplicationManager : ReplicationInstance
-        {
-            public readonly Dictionary<string, ReplicationInstance> Instances;
-
-            public ReplicationManager(string databaseName) : base(databaseName){}
-
-            public ReplicationManager(string databaseName, Dictionary<string, ReplicationInstance> instances) : base(databaseName)
-            {
-                Instances = instances;
-            }
-
-            public override void Break()
-            {
-                foreach (var (node, replicationInstance) in Instances)
-                {
-                    replicationInstance.Break();
-                }
-            }
-
-            public override void Mend()
-            {
-                foreach (var (node, replicationInstance) in Instances)
-                {
-                    replicationInstance.Mend();
-                }
-            }
-
-            public override void ReplicateOnce(string docId)
-            {
-                foreach (var (node, replicationInstance) in Instances)
-                {
-                    replicationInstance.ReplicateOnce(docId);
-                }
-            }
-
-            public override async Task EnsureNoReplicationLoopAsync()
-            {
-                foreach (var (node, replicationInstance) in Instances)
-                {
-                    await replicationInstance.EnsureNoReplicationLoopAsync();
-                }
-            }
-
-            public override void Dispose()
-            {
-                foreach (var instance in Instances.Values)
-                {
-                    instance.Dispose();
-                }
-            }
-
-            internal static async ValueTask<ReplicationManager> GetReplicationManagerAsync(List<RavenServer> servers, string databaseName)
-            {
-                Dictionary<string, ReplicationInstance> instances = new();
-                foreach (var server in servers)
-                {
-                    var instance = await ReplicationInstance.GetReplicationInstanceAsync(server, databaseName);
-                    if(instance != null)
-                        instances[server.ServerStore.NodeTag] = instance;
-                }
-                
-                return new ReplicationManager(databaseName, instances);
-            }
-        }
-
-        public async ValueTask<ReplicationManager> GetReplicationManagerAsync(DocumentStore store, string databaseName, RavenDatabaseMode mode, List<RavenServer> servers = null)
+        public async ValueTask<IReplicationManager> GetReplicationManagerAsync(DocumentStore store, string databaseName, RavenDatabaseMode mode, List<RavenServer> servers = null)
         {
             if (mode == RavenDatabaseMode.Single)
                 return await  ReplicationManager.GetReplicationManagerAsync(servers ?? GetServers() ,databaseName);
